+++ test/import-test-error-no-instances/steps/steps.ts
@@ -1,8 +1,5 @@
 import { expect } from '@playwright/test';
-import { createBdd } from 'playwright-bdd';
-import { test } from './fixtures';
-
-const { Given } = createBdd(test);
+import { Given } from './fixtures';
 
 Given('step 1', async ({ option1 }) => {
   expect(option1).toEqual('foo');