+++ test/import-test-error-no-instances/test.mjs
@@ -5,7 +5,10 @@ const testDir = new TestDir(import.meta);
 test(testDir.name, () => {
   execPlaywrightTestWithError(
     testDir.name,
-    [`Can't guess test instance for: sample.feature`, `Your tests use custom test instance`],
+    [
+      `Can't guess test instance for: sample.feature`, // prettier-ignore
+      `Your tests use custom test instance`,
+    ],
     { cmd: BDDGEN_CMD },
   );
 });